J Instruction
Wow, grueling amount of work just to realize that it wasn't working because I didn't have a wire properly defined.

Added a shift left module for a 26 bit input (to handle the immediate for jump) and a concatenation module to concatenate the current PC with the shifted offset. Then instantiated a mux that chose between the current PC and the new PC given the jump signal being high or not.