BNE
Wow...

Added a new control signal in the control (BNE) and modified the always block to assign this to 0 for all instructions but BNE.

Added the BNE wire to the CPU and hooked it up accordingly. Also changed the logic of PCSrc to go from an AND to a behavioral operation since we want to branch if zero_flag ^ BNE.

Added not_zero to the ALU as well to keep track of when its BNE or BEQ.

Finally, I was having so much trouble with these testbenches so I have added screenshots of each of the test benches for when we demo this.