LUI
Added a concat module thatconcats the imm to the upper 16 bits of the output and wired it accordingly.

Added instruction into control for LUI

all that is not working is that its not writing to the correct register and its not writing the correct value which im not too sure why anymore. going to go into office hours to figure this out.